Backend-llm/requirements.txt
fastapi
uvicorn
python-dotenv
pytest
openai
google-generativeai
groq